Move EvidenceVerdict to entities
